LIB-39 ref:init role sign-up
