+++ src/main/java/com/liberty52/auth/global/exception/external/AuthErrorCode.java
@@ -1,7 +1,6 @@
-package com.liberty52.auth.global.exception;
+package com.liberty52.auth.global.exception.external;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Getter
