+++ src/main/java/com/liberty52/auth/global/exception/external/AuthWithSuchEmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.liberty52.auth.global.exception.external;
+
+public class AuthWithSuchEmailAlreadyExistsException extends AbstractApiException {
+    public AuthWithSuchEmailAlreadyExistsException() {
+        super(AuthErrorCode.EMAIL_ALREADY_EXISTS);
+    }
+}
