+++ src/main/java/com/liberty52/auth/global/exception/external/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.liberty52.auth.global.exception;
+package com.liberty52.auth.global.exception.external;
 
 import org.springframework.http.HttpStatus;
