--- src/main/java/com/liberty52/auth/global/exception/external/Sample.java
@@ -1,4 +1,0 @@
-package com.liberty52.auth.global.exception.external;
-
-public class Sample {
-}
