+++ src/main/java/com/liberty52/auth/global/exception/internal/Sample.java
@@ -1,0 +1,4 @@
+package com.liberty52.auth.global.exception.internal;
+
+public class Sample {
+}
